server/requirements.txt
flask
flask_restful
google-api-python-client
nltk
spacy
pandas
numpy
google_auth_oauthlib
google_auth